Reorder DNSSEC code
